+++ web-site/admin/sql/HISTORY.md
@@ -12,7 +12,10 @@ CREATE TABLE IF NOT EXISTS PASSE_COQUE_MEMBERS (
     AMOUNT INT,
     TELEPHONE VARCHAR(32),
     FIRST_ENROLLED TIMESTAMP,
-    NEWS_LETTER_OK BOOLEAN DEFAULT TRUE
+    NEWS_LETTER_OK BOOLEAN DEFAULT TRUE,
+    PASSWORD VARCHAR(64),
+    ADMIN_PRIVILEGES BOOLEAN DEFAULT FALSE,
+    SOME_CONTENT VARCHAR(512) COMMENT 'Whatever you want goes here'
 );
 
 CREATE TABLE IF NOT EXISTS MEMBER_FEES (
@@ -22,8 +25,8 @@ CREATE TABLE IF NOT EXISTS MEMBER_FEES (
     CONSTRAINT FEES_FK_MEMBERS FOREIGN KEY (EMAIL) REFERENCES PASSE_COQUE_MEMBERS (EMAIL) ON DELETE CASCADE
 );
 ```
-## TODO
-- Add `password`, `admin`, `comments`.
+## An Admin user is added at the end
+- With `password`, `admin`, `comments`.
 
 Le traitement du fichier Excel g&eacute;n&egrave;re un script comme [celui-ci](./create.pc.members.v3.sql).  
 L'en t&ecirc;te du ficher rel&egrave;ve des doublons dans les adresses email.